Moved BarMapper to its own class
