+++ webapp/src/main/java/edu/oakland/cse480/service/AvailableDrinksService.java
@@ -96,28 +96,7 @@ public class AvailableDrinksService extends AbstractJdbcDriver {
             return a;
         }
     }
-    private class BarMapper implements RowMapper<Bar>{
-        public Bar mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Bar r = new Bar();
-            r.setName(rs.getString("name"));
-            r.setId(rs.getInt("id"));
-            r.setBusinessId(rs.getInt("business_id"));
-            r.setOwnerId(rs.getInt("owner_id"));
-            r.setAddress(rs.getString("address"));
-            r.setCity(rs.getString("city"));
-            r.setZipcode(rs.getString("zipcode"));
-            r.setState(rs.getString("state"));
-            r.setNumber(rs.getString("phonenumber"));
-            r.setMondayHours(rs.getString("monday"));
-            r.setTuesdayHours(rs.getString("tuesday"));
-            r.setWednesdayHours(rs.getString("wednesday"));
-            r.setThursdayHours(rs.getString("thursday"));
-            r.setFridayHours(rs.getString("friday"));
-            r.setSaturdayHours(rs.getString("saturday"));
-            r.setSundayHours(rs.getString("sunday"));
-            return r;
-        }
-    }
+
     private class DrinkMapper implements RowMapper<Drink> {
         public Drink mapRow(ResultSet rs, int rowNum) throws SQLException {
             Drink d = new Drink();
